Update SenseHat project
